Add `plugins` module
