+++ src/felupe/__init__.py
@@ -74,7 +74,6 @@ from .field import (
 )
 from .mechanics import (
     CharacteristicCurve,
-    CharacteristicCurvePlugin,
     ContactRigidPlane,
     FormItem,
     FreeVibration,
@@ -92,6 +91,7 @@ from .mechanics import (
     TrussBody,
 )
 from .mesh import Circle, Cube, Grid, Mesh, MeshContainer, Point, Rectangle
+from .plugins import CharacteristicCurvePlugin, ProgressPlugin
 from .quadrature import (
     BazantOh,
     GaussLegendre,
@@ -129,7 +129,6 @@ from .region import (
     RegionVertex,
 )
 from .tools import (
-    ProgressPlugin,
     hello_world,
     newtonraphson,
     newtonrhapson,
@@ -222,8 +221,6 @@ __all__ = [
     "Point",
     "Rectangle",
     "CharacteristicCurve",
-    "CharacteristicCurvePlugin",
-    "ProgressPlugin",
     "Job",
     "PointLoad",
     "ContactRigidPlane",
