+++ src/felupe/mechanics/__init__.py
@@ -14,12 +14,10 @@ from ._solidbody_pressure import SolidBodyPressure
 from ._step import Step
 from ._truss import TrussBody
 from ._update import UpdateItem
-from .plugins import CharacteristicCurvePlugin
 
 __all__ = [
     "Assemble",
     "CharacteristicCurve",
-    "CharacteristicCurvePlugin",
     "Evaluate",
     "FreeVibration",
     "FormItem",
