+++ src/felupe/plugins/__init__.py
@@ -1,5 +1,7 @@
 from ._characteristic_curve import CharacteristicCurvePlugin
+from ._progress import ProgressPlugin
 
 __all__ = [
     "CharacteristicCurvePlugin",
+    "ProgressPlugin",
 ]
