+++ src/felupe/tools/_event_dispatcher.py
@@ -34,33 +34,10 @@ HOOKS = (
 class Context:
     "A class to keep track of the context of a Job during evaluation."
 
-    def __init__(self, job=None, step=None, substep=None, newton=None):
+    def __init__(self, job=None, step=None, substep=None):
         self.job = job
         self.step = step
         self.substep = substep
-        self.newton = newton
-
-
-class JobState:
-    "A class to keep track of the state of a Job during evaluation."
-
-    def __init__(
-        self,
-        stepnumber=None,
-        substepnumber=None,
-        iteration=None,
-        fnorm=None,
-        xnorm=None,
-        tol=None,
-        success=None,
-    ):
-        self.stepnumber = stepnumber
-        self.substepnumber = substepnumber
-        self.iteration = iteration
-        self.fnorm = fnorm
-        self.xnorm = xnorm
-        self.success = success
-        self.tol = tol
 
 
 class EventDispatcher:
